Block Request from different organization, for OrganizationAdmin and …

+++ app/Http/Controllers/OrganizationAdminController.php
@@ -31,14 +31,15 @@ class OrganizationAdminController extends Controller
      */
     public function index()
     {
-        $result = User::where('role_id', $this->organization_admin_role->id)
-            ->get()
-            ->toArray();
+        $organization_id = auth()->user()->organization_id;
+        $organization_admins = User::where('organization_id', $organization_id)
+            ->where('role_id', $this->organization_admin_role->id)
+            ->get();
 
         return response()->json(
             [
                 'message' => 'Organization Admin List',
-                'data' => $result,
+                'data' => $organization_admins,
             ],
             Response::HTTP_OK
         );
@@ -52,13 +53,16 @@ class OrganizationAdminController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $organization_id = auth()->user()->organization_id;
+
         $user = User::create([
             'username' => $request->username,
             'email' => $request->email,
             'first_name' => $request->first_name,
             'last_name' => $request->last_name,
-            'password' => Hash::make($request->password),
             'role_id' => $this->organization_admin_role->id,
+            'organization_id' => $organization_id,
+            'password' => Hash::make($request->password),
         ]);
 
         return response()->json(
@@ -79,12 +83,25 @@ class OrganizationAdminController extends Controller
      */
     public function update(AdminRequest $request, User $user)
     {
+        $organization_id = auth()->user()->organization_id;
+
+        if ($user->organization_id != $organization_id) {
+            return response()->json(
+                [
+                    'message' =>
+                        'Could not update an admin in different Organization',
+                ],
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
         $user->update([
             'username' => $request->username,
             'email' => $request->email,
             'first_name' => $request->first_name,
             'last_name' => $request->last_name,
             'role_id' => $this->organization_admin_role->id,
+            'organization_id' => $organization_id,
         ]);
 
         return response()->json(
@@ -104,6 +121,18 @@ class OrganizationAdminController extends Controller
      */
     public function destroy(User $user)
     {
+        $organization_id = auth()->user()->organization_id;
+
+        if ($user->organization_id != $organization_id) {
+            return response()->json(
+                [
+                    'message' =>
+                        'Could not remove an admin in different Organization',
+                ],
+                Response::HTTP_FORBIDDEN
+            );
+        }
+
         $user->delete();
 
         return response()->json(
